♻️ refactor: 패키지 구조 변경
- MVC 패턴 적용
