+++ src/main/java/racingcar/Application.java
@@ -1,5 +1,9 @@
 package racingcar;
 
+import racingcar.controller.Game;
+import racingcar.view.Input;
+import racingcar.view.Output;
+
 public class Application {
     public static void main(String[] args) {
         // TODO: 프로그램 구현
