+++ src/main/java/racingcar/controller/Race.java
@@ -1,7 +1,8 @@
-package racingcar;
+package racingcar.controller;
 
 import java.util.Arrays;
 import java.util.List;
+import racingcar.model.Car;
 
 public class Race {
