+++ src/main/java/racingcar/view/Output.java
@@ -1,7 +1,8 @@
-package racingcar;
+package racingcar.view;
 
 import java.util.List;
 import java.util.stream.Collectors;
+import racingcar.model.Car;
 
 public class Output {
